refactor: extract FileDiffLegend to dedicated file
AnalysisDetailPanel.tsx 内で定義されていた FileDiffLegend を
独立ファイル（FileDiffLegend.tsx）に抽出。
Stories からも参照されており、見通しの改善のため分離した。
