+++ frontend/src/components/AnalysisDetailPanel.tsx
@@ -10,6 +10,7 @@ import type {
 import { Card, Panel } from "./Card";
 import { Badge } from "./Badge";
 import { RiskBadge } from "./RiskBadge";
+import { FileDiffLegend } from "./FileDiffLegend";
 
 interface AnalysisDetailPanelProps {
   result: AnalysisResult;
@@ -156,24 +157,6 @@ function RiskFactorList({ factors }: { factors: RiskFactor[] }) {
           </span>
         </div>
       ))}
-    </div>
-  );
-}
-
-export function FileDiffLegend() {
-  const { t } = useTranslation();
-
-  return (
-    <div className="flex items-center gap-3 text-[0.75rem] text-text-secondary">
-      <span>{t("pr.fileDiffLegend")}</span>
-      <span className="flex items-center gap-1">
-        <span className="inline-block h-2 w-2 rounded-full bg-accent" />
-        <span>+ {t("pr.fileDiffLegendAdditions")}</span>
-      </span>
-      <span className="flex items-center gap-1">
-        <span className="inline-block h-2 w-2 rounded-full bg-danger" />
-        <span>- {t("pr.fileDiffLegendDeletions")}</span>
-      </span>
     </div>
   );
 }
